requirements.txt
# AI / LLMs
langchain==0.1.12
openai==1.14.1
llama-index

# Backend
fastapi==0.109.2
uvicorn==0.27.0.post1   # ASGI server for FastAPI

# Database
sqlalchemy==2.0.27
sqlite-utils==3.35
alembic==1.13.1         # DB migrations (optional but useful)

# Environment Variables
python-dotenv==1.0.1

# Logging & Scheduling
loguru==0.7.2           # better logging
apscheduler==3.10.4     # scheduling tasks (e.g., reminders)

# CLI Tools
typer==0.12.3           # command-line app support

# Web App / Dashboard
streamlit==1.35.0       # UI for your water tracker

# HTTP Clients
requests==2.31.0
httpx==0.27.0           # async-friendly client

# Validation
pydantic==2.6.1

# Dev Tools (optional but recommended)
black==24.3.0
isort==5.13.2

